// src/model/vote.ts
export interface vote {
    E:          number;
    sA:         number;
    sB:         number;
    K:          number;
    S:          number;
    score:      number;
    ss:         number;
    did:        number;
}

const defaultVote: vote = {
    E: 0,
    sA: 0,
    sB: 0,
    K: 0,
    S: 0,
    score: 0,
    ss: 0,
    did: 0
};